implemented the change of step sequence in seller registration
